Part4 d

+++ bloglist-app/controllers/blogs.js
@@ -1,26 +1,58 @@
-const blogsRouter = require("express").Router();
-const Blog = require("../models/blog");
+const jwt = require('jsonwebtoken')
+const blogsRouter = require('express').Router()
+const Blog = require('../models/blog')
+const User = require('../models/user')
 
-blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({})
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', {
+    username: 1,
+    name: 1,
+    id: 1
+  })
   response.json(blogs)
-});
+})
 
-blogsRouter.post("/", async (request, response, next) => {
-  const blog = new Blog({
-    ...request.body,
-    likes: request.body.likes === undefined ? 0 : request.body.likes
-  });
-  
+blogsRouter.get('/:id', async (request, response, next) => {
   try {
-    const savedBlogs = await blog.save()
-    response.status(201).json(savedBlogs)
+    const blog = await Blog.findById(request.params.id)
+    if (blog) {
+      response.json(blog.toJSON())
+    } else {
+      response.status(404).end()
+    }
   } catch (exception) {
     next(exception)
   }
-});
+})
 
-blogsRouter.put("/:id", async (request, response, next) => {
+blogsRouter.post('/', async (request, response, next) => {
+  const body = request.body
+
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = new Blog({
+      ...body,
+      likes: body.likes === undefined ? 0 : body.likes,
+      user
+    })
+
+    const savedBlogs = await blog.save()
+    user.blogs = user.blogs.concat(savedBlogs._id)
+    await user.save()
+
+    response.status(201).json(savedBlogs.toJSON())
+  } catch (exception) {
+    next(exception)
+  }
+})
+
+blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
 
   const blog = {
@@ -29,20 +61,35 @@ blogsRouter.put("/:id", async (request, response, next) => {
   }
 
   try {
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
+      new: true
+    })
     response.json(updatedBlog.toJSON())
   } catch (exception) {
     next(exception)
   }
 })
 
-blogsRouter.delete("/:id", async (request, response, next) => {
+blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+  
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
+
+    return response.status(401).json({ error: 'invalid token' })
+
   } catch (exception) {
     next(exception)
   }
-});
+})
 
-module.exports = blogsRouter;
+module.exports = blogsRouter
